Really messy cleanup.  This reverts part of #16640 - reintroducing codeql_config and the usage of it in the github workflow.  DO NOT MERGE THIS INTO MAIN.
Keeping this old branch around as a reminder of hubris and the inexorable weight of icarus while I see if it can be turned into something else.
I'm forced to make a commit before branches are cleaned up, so here we are.

+++ pkg/security/secl/model/consts_linux_amd64.go
@@ -4,8 +4,8 @@
 // Copyright 2016-present Datadog, Inc.
 
 //go:build linux
-// +build linux
 
+// Package model holds model related files
 package model
 
 import (
@@ -13,6 +13,8 @@ import (
 )
 
 var (
+	// ptraceArchConstants are the supported ptrace commands for the ptrace syscall on amd64
+	// generate_constants:Ptrace constants,Ptrace constants are the supported ptrace commands for the ptrace syscall.
 	ptraceArchConstants = map[string]uint32{
 		"PTRACE_GETFPREGS":         unix.PTRACE_GETFPREGS,
 		"PTRACE_SETFPREGS":         unix.PTRACE_SETFPREGS,
@@ -27,7 +29,9 @@ var (
 		"PTRACE_SINGLEBLOCK":       unix.PTRACE_SINGLEBLOCK,
 	}
 
-	mmapFlagArchConstants = map[string]int{
+	// mmapFlagConstants are the supported flags for the mmap syscall on amd64
+	// generate_constants:MMap flags,MMap flags are the supported flags for the mmap syscall.
+	mmapFlagArchConstants = map[string]uint64{
 		"MAP_32BIT": unix.MAP_32BIT, /* only give out 32bit addresses */
 	}
 )
