+++ pkg/security/secl/model/consts_linux_arm64.go
@@ -4,8 +4,8 @@
 // Copyright 2016-present Datadog, Inc.
 
 //go:build linux
-// +build linux
 
+// Package model holds model related files
 package model
 
 import (
@@ -13,6 +13,8 @@ import (
 )
 
 var (
+	// ptraceArchConstants are the supported ptrace commands for the ptrace syscall on arm64
+	// generate_constants:Ptrace constants,Ptrace constants are the supported ptrace commands for the ptrace syscall.
 	ptraceArchConstants = map[string]uint32{
 		"PTRACE_PEEKMTETAGS":       unix.PTRACE_PEEKMTETAGS,
 		"PTRACE_POKEMTETAGS":       unix.PTRACE_POKEMTETAGS,
@@ -20,5 +22,5 @@ var (
 		"PTRACE_SYSEMU_SINGLESTEP": unix.PTRACE_SYSEMU_SINGLESTEP,
 	}
 
-	mmapFlagArchConstants = map[string]int{}
+	mmapFlagArchConstants = map[string]uint64{}
 )
